mgmtd: step 4: FE adapter get-tree functionality

+++ lib/mgmt_fe_client.h
@@ -119,7 +119,7 @@ struct mgmt_fe_client_cbs {
 	/* Called when get-tree result is returned */
 	int (*get_tree_notify)(struct mgmt_fe_client *client,
 			       uintptr_t user_data, uint64_t client_id,
-			       uint64_t session_id, uint64_t session_ctx,
+			       uint64_t session_id, uintptr_t session_ctx,
 			       uint64_t req_id, Mgmtd__DatastoreId ds_id,
 			       LYD_FORMAT result_type, void *result, size_t len,
 			       int partial_error);
